+++ Changelog.md
@@ -21,6 +21,7 @@ Bugfixes:
 Build System:
  * Add support for continuous fuzzing via Google oss-fuzz
  * Ubuntu PPA Packages: Use CVC4 as SMT solver instead of Z3
+ * Soltest: Add parser that is used in the file-based unit test environment.
 
 
 ### 0.5.3 (2019-01-22)